feat(e2e): split up e2e into more jobs
