+++ crates/e2e/mise.toml
@@ -88,6 +88,14 @@ docker exec "$NODE" sh -c '
   mkdir -p /kopiur-e2e/repo /kopiur-e2e/src/sub
   printf "hello kopiur e2e\n" > /kopiur-e2e/src/a.txt
   printf "nested data\n"      > /kopiur-e2e/src/sub/b.txt
+  # Per-scenario ISOLATED repo dirs (ADR-0004/0005 scenarios). The operator mounts a
+  # filesystem repo PVC at backend.path AND runs kopia --path=backend.path, so the
+  # PVC ROOT is the kopia repo — a subdir under one shared PVC gives no isolation.
+  # Each scenario binds its own PVC over one of these dirs. KEEP IN LOCKSTEP with
+  # crates/e2e/src/consts.rs::REPO_SUBPATHS.
+  for s in moverdefaults nsdel-orphan nsdel-delete pin populator readonly kstatus verify repl-src repl-dst projgate; do
+    mkdir -p "/kopiur-e2e/repos/$s"
+  done
   chmod -R 0777 /kopiur-e2e
   # A deliberately NON-writable repo dir (root-owned, no write bit) for the
   # terminal-failure regression test; set AFTER the 0777 above so it sticks.
