+++ deploy/helm/kopiur/files/crds/snapshotpolicies.yaml
@@ -10480,13 +10480,15 @@ spec:
                   hostname:
                     description: |-
                       Override the `hostname` portion of `username@hostname:path`; absent uses the
-                      resolved default. Templated with `tera` and pinned at admission (ADR §4.2).
+                      resolved default (the repository's `identityDefaults` CEL expression, or the
+                      namespace). Used verbatim and pinned at admission (ADR §4.2).
                     nullable: true
                     type: string
                   username:
                     description: |-
                       Override the `username` portion of `username@hostname:path`; absent uses the
-                      resolved default. Templated with `tera` and pinned at admission (ADR §4.2).
+                      resolved default (the repository's `identityDefaults` CEL expression, or the
+                      object name). Used verbatim and pinned at admission (ADR §4.2).
                     nullable: true
                     type: string
                 type: object
